Now CORRECTLY handles the main bug without breaking the integration tests. Probably a little gross but doesn't break anything anymore
