+++ Development/Source/tapestry.cfg
@@ -2,7 +2,7 @@
 uid = patches
 compid = Testbed
 blocksize = 4096
-expected fp = 4961A99F45B8CDFBB0D25745C746293D20FCAC20
+expected fp = 9997AC649C3A2F2A4CBC158B67CB85C9D16755C4
 sign by default = True
 signing fp = 388BB3657EAB090A9C01EC12288CBA1D80DCB8A5
 recovery path = /home/patches/Desktop/Tapestry FT Data/Test/Inc
