+++ Development/Source/tapestry/__init__.py
@@ -1,3 +1,4 @@
 __version__ = "2.0.2"
 
 from .classes import *
+from .functions import *
